+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [v1.0.1] - 10.08.2024
+
+## Added
+
+- When local player has taken damage or healed, amount of health that was changed will pop up on screen instead of world space
+
+## Fixed
+
+- Damage popups not cleaning up after fully fading out
+
+
 # [v1.0.0] - 08.08.2024
 
 ## Added